Simpler solution to the multiple command-on call problem.
We simply remember the fullframe state in a buffer local variable only
set and unset in the target buffer.  If it is already set, we don't store
the current window configuration.

This indeed fixes #13 for me with the usage of fullframe in
`repl-toggl`.  The magit scenario as describe by @purcell in #15 is
reproducible even when no fullframe.el advice to magit is present.