Only generate code for the current version of Emacs
This saves the conditional definition of old/new advice code, with the
trade-off that in the unlikely event someone byte-compiles code using
`fullframe' on Emacs >= 24.4 and then runs it on an older Emacs, they
will receive an error that advice-add is not defined. In the opposite
and more likely case, wherein someone upgrades their Emacs and still has
old compiled byte code, it will continue to work.